Overly permissive LUB for qualified attr types to match Rust
The least upper bound between a required and optional qualified
attribute type is defined as an optional qualified type to match the
Rust implementation.  This is sound, but is too permissive for our SMT
encoding. We apply the change the 3.1.x release branch nonetheless so
that differential testing does not report this discrepancy.

Signed-off-by: John Kastner <jkastner@amazon.com>