Add fuzz targets using arbitrary strings
Some of our fuzz target work by taking generating a structured policy,
printing it to a string, and then testing properties of that string.
This approach is limited because it will not test policy strings that
are not generated by our policy printing function. We can avoid this by
doing unstructured fuzzing.

To make this targets find interseting inputs in a reasonable amount of
time, you will need to seed the corpus with some valid cedar/schema
file using `./intialize-corpus.sh --all`

Signed-off-by: John Kastner <jkastner@amazon.com>