Adjust level based entity slicing to be infallible
This procedure would previously return `none` if it dereferenced an
entity which did not exist in the original entities. It now returns an
entity slice without the missing entity. This means that evaluating a
request with the slice may result in an unknown entity error, but the
soundness proofs shows that that this only happens when the error
also occurred using the unsliced entities.

Signed-off-by: John Kastner <jkastner@amazon.com>